# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mister — il fantacalcio manageriale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/federation:
    get:
      operationId: getFederation
      tags: [federations]
      summary: Regolamento della lega
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Federation"
        "404":
          description: Not found
    patch:
      operationId: updateFederation
      tags: [federations]
      summary: Aggiorna regolamento (admin lega)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FederationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Federation"
        "404":
          description: Not found

  # ─── COMPETITIONS ─────────────────────────────────────────────
  /leagues/{leagueId}/competitions:
    get:
      operationId: listCompetitions
      tags: [competitions]
      summary: Lista competizioni della lega
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competition"
    post:
      operationId: createCompetition
      tags: [competitions]
      summary: Crea competizione
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /leagues/{leagueId}/competitions/{id}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Dettaglio competizione
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
        "404":
          description: Not found
    patch:
      operationId: updateCompetition
      tags: [competitions]
      summary: Aggiorna competizione
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
        "404":
          description: Not found
    delete:
      operationId: deleteCompetition
      tags: [competitions]
      summary: Elimina competizione
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "204":
          description: Deleted

  # ─── MARKETS ─────────────────────────────────────────────────
...
  /leagues/{leagueId}/teams:
    get:
      operationId: listFantaTeams
      tags: [teams]
      summary: Lista squadre della lega
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FantaTeam"
    post:
      operationId: createFantaTeam
      tags: [teams]
      summary: Crea squadra
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FantaTeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FantaTeam"
  /leagues/{leagueId}/teams/{id}:
    get:
      operationId: getFantaTeam
      tags: [teams]
      summary: Dettaglio squadra
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FantaTeam"
        "404":
          description: Not found
    patch:
      operationId: updateFantaTeam
      tags: [teams]
      summary: Aggiorna squadra
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FantaTeamUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FantaTeam"
        "404":
          description: Not found

  # ─── PLAYERS ─────────────────────────────────────────────────
...
  /leagues/{leagueId}/contracts:
    get:
      operationId: getLeagueContracts
      tags: [contracts]
      summary: Lista contratti della lega
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Crea contratto (post-asta)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /leagues/{leagueId}/contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Dettaglio contratto
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Aggiorna stato contratto (rescissione, clausola pagata, ecc.)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found

  # ─── DASHBOARD (wow endpoint) ─────────────────────────────────
...
components:
  parameters:
    StringId:
      in: path
      name: id
      required: true
      schema:
        type: string
    LeagueId:
      in: path
      name: leagueId
      required: true
      schema:
        type: string
  schemas:
...
    FeatureFlags:
      type: object
      description: I 18 feature flag copiati dal template al momento della creazione della lega
      properties:
        multi_season_contracts:
          type: boolean
          description: Contratti pluriennali (>1 stagione)
        max_contract_length:
          type: integer
          description: Durata massima contratto in stagioni (1-10)
        contract_renewal:
          type: boolean
          description: Flusso esplicito di rinnovo contrattuale
        preemption_right:
          type: boolean
          description: Diritto di pareggio in asta per il detentore uscente
        repair_auction_january:
          type: boolean
          description: Asta di riparazione invernale di gennaio
        free_agent_pool:
          type: boolean
          description: Pool svincolati permanente disponibile
        direct_trades:
          type: boolean
          description: Scambi diretti 1-a-1 tra manager
        always_on_markets:
          type: boolean
          description: Mercati e scambi sempre attivi in stagione
        carryover_budget:
          type: boolean
          description: Budget residuo portato alla stagione successiva
        carryover_percentage:
          type: number
          description: Percentuale di budget residuo mantenuta (0-100)
        player_value_dynamic:
          type: boolean
          description: Valore di mercato del giocatore aggiornato dinamicamente
        amortization:
          type: boolean
          description: Ammortamento del prezzo d'acquisto sugli anni di contratto
        release_clauses:
          type: boolean
          description: Clausole rescissorie attive
        clause_default_factor:
          type: number
          description: Moltiplicatore residuo per clausola di default (0.5-1.0)
        rescission_penalty:
          type: boolean
          description: Penale se il manager svincola prima della scadenza
        rescission_recovery_pct:
          type: number
          description: Percentuale crediti recuperata sullo svincolo (0-100)
        no_schema_tactics:
          type: boolean
          description: Formazione senza vincoli di modulo predefinito
        scouting_enabled:
          type: boolean
          description: Sistema di scouting giocatori giovani/emergenti

    # ─── MARKET EVENT SUGGESTION (nel template) ───────────────
...
    FederationRules:
      type: object
      description: Regole di calcolo del punteggio fanta (bonus/malus, soglie, modificatori)
      properties:
        bonusMalus:
          type: object
          description: Bonus e malus per evento (gol, assist, ammonizioni, ecc.)
        goalThresholds:
          type: object
          description: Conversione punteggio squadra in fanta-gol
        defenseModifier:
          type: object
          description: Modificatore difesa
        midfieldModifier:
          type: object
          description: Modificatore centrocampo
        homeAdvantage:
          type: object
          description: Bonus padrone di casa
        substitutions:
          type: object
          description: Regole sostituzioni automatiche

    # ─── FEDERATION ───────────────────────────────────────────
    Federation:
      type: object
      required: [id, name, mode, feature_flags, rules, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        template_id:
          type: ["string", "null"]
        mode:
          type: string
          enum: [classic, mantra]
        feature_flags:
          $ref: "#/components/schemas/FeatureFlags"
        rules:
          $ref: "#/components/schemas/FederationRules"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FederationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        mode:
          type: string
          enum: [classic, mantra]
        feature_flags:
          $ref: "#/components/schemas/FeatureFlags"
        rules:
          $ref: "#/components/schemas/FederationRules"

    # ─── COMPETITION ──────────────────────────────────────────
    Competition:
      type: object
      required: [id, league_id, name, type, season, start_giornata, end_giornata, config, active, completed, created_at]
      properties:
        id:
          type: string
        league_id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [campionato, coppa, battle_royale, sprint_race, formula_uno, punteggio_assoluto]
        season:
          type: integer
        start_giornata:
          type: integer
        end_giornata:
          type: integer
        config:
          type: object
          description: Configurazione specifica per tipo + tiebreaker + premi + partecipanti
        active:
          type: boolean
        completed:
          type: boolean
        starts_at:
          type: ["string", "null"]
          format: date-time
        ends_at:
          type: ["string", "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
    CompetitionInput:
      type: object
      required: [name, type, season, start_giornata, end_giornata]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [campionato, coppa, battle_royale, sprint_race, formula_uno, punteggio_assoluto]
        season:
          type: integer
        start_giornata:
          type: integer
        end_giornata:
          type: integer
        participant_team_ids:
          type: array
          items:
            type: string
        tiebreakers:
          type: array
          items:
            type: string
        settings:
          type: object
    CompetitionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        start_giornata:
          type: integer
        end_giornata:
          type: integer
        settings:
          type: object
        active:
          type: boolean
        completed:
          type: boolean

    # ─── MARKET EVENT ─────────────────────────────────────────
...
    FantaTeam:
      type: object
      required: [id, league_id, manager_user_id, name, credits_remaining, created_at]
      properties:
        id:
          type: string
        league_id:
          type: string
        manager_user_id:
          type: string
        name:
          type: string
        name_auction:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        credits_remaining:
          type: integer
        roster:
          type: array
          items:
            type: integer
        created_at:
          type: string
          format: date-time
    FantaTeamInput:
      type: object
      required: [manager_user_id, name]
      properties:
        manager_user_id:
          type: string
        name:
          type: string
        name_auction:
          type: string
        logo_url:
          type: string
    FantaTeamUpdate:
      type: object
      properties:
        name:
          type: string
        name_auction:
          type: string
        logo_url:
          type: string
        credits_remaining:
          type: integer
        roster:
          type: array
          items:
            type: integer

    # ─── PLAYER ───────────────────────────────────────────────
...
    Contract:
      type: object
      required: [id, league_id, fanta_team_id, player_id, season_start, duration_seasons, purchase_price, clause_default, clause_investment, state, created_at]
      properties:
        id:
          type: string
        league_id:
          type: string
        fanta_team_id:
          type: string
        player_id:
          type: integer
        season_start:
          type: integer
        duration_seasons:
          type: integer
        purchase_price:
          type: integer
        clause_default:
          type: integer
        clause_investment:
          type: integer
        state:
          type: string
          enum: [active, expired, rescinded, clause_paid, traded]
        notes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        closed_at:
          type: ["string", "null"]
          format: date-time
    ContractInput:
      type: object
      required: [fanta_team_id, player_id, season_start, duration_seasons, purchase_price, clause_default]
      properties:
        fanta_team_id:
          type: string
        player_id:
          type: integer
        season_start:
          type: integer
        duration_seasons:
          type: integer
        purchase_price:
          type: integer
        clause_default:
          type: integer
        clause_investment:
          type: integer
    ContractUpdate:
      type: object
      properties:
        state:
          type: string
          enum: [active, expired, rescinded, clause_paid, traded]
        clause_investment:
          type: integer
        notes:
          type: array
          items:
            type: string
        closed_at:
          type: string
          format: date-time

    # ─── DASHBOARD ────────────────────────────────────────────
...