# lib/api-zod/src/generated/types/competition.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mister — il fantacalcio manageriale
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitionConfig } from './competitionConfig';
import type { CompetitionType } from './competitionType';

export interface Competition {
  id: string;
  league_id: string;
  name: string;
  description?: string;
  type: CompetitionType;
  season: number;
  start_giornata: number;
  end_giornata: number;
  /** Configurazione specifica per tipo + tiebreaker + premi + partecipanti */
  config: CompetitionConfig;
  active: boolean;
  completed: boolean;
  /** @nullable */
  starts_at?: Date | null;
  /** @nullable */
  ends_at?: Date | null;
  created_at: Date;
}