# lib/api-zod/src/generated/types/competitionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mister — il fantacalcio manageriale
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitionUpdateSettings } from './competitionUpdateSettings';

export interface CompetitionUpdate {
  name?: string;
  description?: string;
  start_giornata?: number;
  end_giornata?: number;
  settings?: CompetitionUpdateSettings;
  active?: boolean;
  completed?: boolean;
}