# lib/api-zod/src/generated/types/federation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mister — il fantacalcio manageriale
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureFlags } from './featureFlags';
import type { FederationMode } from './federationMode';
import type { FederationRules } from './federationRules';

export interface Federation {
  id: string;
  name: string;
  description?: string;
  /** @nullable */
  template_id?: string | null;
  mode: FederationMode;
  feature_flags: FeatureFlags;
  rules: FederationRules;
  created_at: Date;
  updated_at: Date;
}